Process on multiple cores on one machine

This makes use of rzmq messaging and sends requests via TCP/IP

Super class

clustermq::QSys -> MULTICORE

Methods

Public methods

MULTICORE$new()

MULTICORE$submit_jobs()

MULTICORE$cleanup()

MULTICORE$finalize()

MULTICORE$clone()

Inherited methods

clustermq::QSys$receive_data()

clustermq::QSys$send_call()

clustermq::QSys$send_common_data()

clustermq::QSys$send_job_data()

clustermq::QSys$send_shutdown_worker()

clustermq::QSys$send_wait()

clustermq::QSys$set_common_data()

Method new()

Usage

MULTICORE$new(...)

Method submit_jobs()

Usage

MULTICORE$submit_jobs(n_jobs, ...)

Method cleanup()

Usage

MULTICORE$cleanup(quiet = FALSE, timeout = 3)

Method finalize()

Usage

MULTICORE$finalize(quiet = FALSE)

Method clone()

The objects of this class are cloneable with this method.

Usage

MULTICORE$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.