Wrap C++ Rcpp module in R6 to get reliable argument matching

This is an R6 wrapper of the C++ class in order to support R argument matching. Ideally, Rcpp will at some point support this natively and this file will no longer be necessary. Until then, it causes redundancy with zeromq.cpp, but this is a small inconvenience and much less error-prone than only relying on positional arguments.

Methods

Public methods

ZeroMQ$new()

ZeroMQ$listen()

ZeroMQ$connect()

ZeroMQ$disconnect()

ZeroMQ$send()

ZeroMQ$receive()

ZeroMQ$poll()

Method new()

Usage

ZeroMQ$new()

Method listen()

Usage

ZeroMQ$listen(addrs = host(), socket_type = "ZMQ_REP", sid = "default")

Method connect()

Usage

ZeroMQ$connect(address, socket_type = "ZMQ_REQ", sid = "default")

Method disconnect()

Usage

ZeroMQ$disconnect(sid = "default")

Method send()

Usage

ZeroMQ$send(data, sid = "default", dont_wait = FALSE, send_more = FALSE)

Method receive()

Usage

ZeroMQ$receive(sid = "default", dont_wait = FALSE, unserialize = TRUE)

Method poll()

Usage

ZeroMQ$poll(sid = "default", timeout = -1L)