Fill a template string with supplied values

Fill a template string with supplied values

Usage

fill_template(template, values, required = c())

Arguments

template

A character string of a submission template

values

A named list of key-value pairs

required

Keys that must be present in the template (default: none)

Value

A template where placeholder fields were replaced by values