Class for basic queuing system functions

Provides the basic functions needed to communicate between machines This should abstract most functions of rZMQ so the scheduler implementations can rely on the higher level functionality

Methods

Public methods

QSys$new()

QSys$submit_jobs()

QSys$send_call()

QSys$set_common_data()

QSys$send_common_data()

QSys$send_job_data()

QSys$send_wait()

QSys$receive_data()

QSys$send_shutdown_worker()

QSys$cleanup()

Method new()

Usage

QSys$new(
  addr = host(),
  bind = TRUE,
  data = NULL,
  reuse = FALSE,
  template = NULL,
  zmq = ZeroMQ$new()
)

Method submit_jobs()

Usage

QSys$submit_jobs(...)

Method send_call()

Usage

QSys$send_call(expr, env = list(), ref = substitute(expr))

Method set_common_data()

Usage

QSys$set_common_data(...)

Method send_common_data()

Usage

QSys$send_common_data()

Method send_job_data()

Usage

QSys$send_job_data(...)

Method send_wait()

Usage

QSys$send_wait(wait = 0.05 * self$workers_running)

Method receive_data()

Usage

QSys$receive_data(timeout = Inf, with_checks = TRUE)

Method send_shutdown_worker()

Usage

QSys$send_shutdown_worker()

Method cleanup()

Usage

QSys$cleanup(quiet = FALSE, timeout = 5)